surface product slug for client-side consumption.
In order for the persona bar to have product context when on a product page, we'll need to surface the product slug, in a way that it can be accessed client side. adding a hidden filed with the slug value will allow the persona bar to detect a product page and use the value to obtain product details via a webservice call.