Adding missing Xcillion skin
I was setting up a new dev environment for HCC and got my database setup according to the docs. When I went to run the development website I this DNN error

[D:2][T:9][ERROR] DotNetNuke.Services.Exceptions.Exceptions - DotNetNuke.Services.Exceptions.PageLoadException: Unhandled error loading page. ---> System.Web.HttpException: The file '/Portals/_default/Skins/Xcillion/Inner.ascx' does not exist.

Looks like the default Skin was missing from the repo. 

This change will add the missing skin folder to the repo to resolve the above mentioned error when setting up a new dev environment