Check Store ID when Checking for Duplicate Variants
When multiple stores exist and you attempt to add a variant to a product it does not currently filter by StoreId and gives an error message abut the SKU already existing. Even though the SKU doesn't exist in the current store. 

This adds a stored check based on the current context.